% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSchema.R
\name{create_schema}
\alias{create_schema}
\title{Creates database in Postgres}
\usage{
create_schema(tables = c("mtd", "ind_dat", "ind_boro_dat"))
}
\arguments{
\item{tables}{There are also five implicit parameters required to connect to database, currently as environment variables.}
}
\value{
The function creates tables in Postgres database and returns TRUE if the run doesn't result in fatal error
}
\description{
Creates database in Postgres
}
\examples{
create_schema(tables = c("mtd", "ind_dat", "ind_boro_dat"))
}
