% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_boro_db.R
\name{insert_boro_db}
\alias{insert_boro_db}
\title{Add borough-level data to database}
\usage{
insert_boro_db(dat, log = "")
}
\arguments{
\item{dat}{A data.frame with correct columns for the boro table and at least one row}

\item{log}{Location of log file. Default is print to console}
}
\value{
Logical - TRUE for worked ok.
}
\description{
Add borough-level data to database
}
\examples{
dat<-data.frame(dataset='test',x=1, y=9, lad11nm="test")
insert_boro_db(dat)
}
