% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_db.R
\name{insert_db}
\alias{insert_db}
\title{Add data to database}
\usage{
insert_db(dat, log = "")
}
\arguments{
\item{dat}{A data.frame with correct columns and at least one row}

\item{log}{Location of log file. Default is print to console}
}
\value{
Logical - TRUE for worked OK.
}
\description{
Add data to database
}
\examples{
dat <- data.frame(dataset='test',xwhich=0,xvarchar='test',xvardt=1, yval=0, yvllb='test', text='test')
insert_db(dat)
}
